Convert vimscript to lua to a certain extent
convert vimscript config to lua to a certain extent. The vim lsp still
does not work very well for rust. It works pretty well for go.
Autocomplete for rust is broken for my config even with other plugins.
Coc seems to work fine. So the autocomplete part is written in vimscript
and everything else is in lua scripts.